telemetry: integrate messages and drive validation

+++ src/services/main/control_app.h
@@ -73,16 +73,21 @@ inline void app_dispatcher(void* ctx, MsgId id, const void* payload) {
     if constexpr (Config::kPrintEvery != -1) {
       if ((++telemetry_count % Config::kPrintEvery) == 0) {
         const ipc::SystemTelemetryPayload& p = unpack_payload<MsgId::SystemTelemetry>(payload);
-        const bool motor_dt_warning = p.motor_update_dt_ms > (1500.0 / Config::control_hz);
-        const double applied_avg_sps = 0.5 * (p.left_applied_sps + p.right_applied_sps);
+        const bool motor_dt_warning =
+            p.motor_update_dt_ms > (1500.0f / static_cast<float>(Config::control_hz));
+        const float applied_avg_sps = 0.5f * (p.left_applied_sps + p.right_applied_sps);
         std::printf(
             "t=%7.3f  th=%6.2f deg  dth=%7.2f dps  u=%6.0f%s  "
             "sp=%6.2f (%+5.2f/%+5.2f)  perr=%6.2f  v=%7.1f/%7.1f  "
             "ap=%6.0f%s  trn=%6.0f\n",
-            p.t_sec, p.pitch_deg, p.pitch_rate_dps, p.u_sps,
-            (std::abs(p.u_sps) >= 0.99 * kMaxSps) ? "*" : "", p.pitch_sp_deg,
-            p.velocity_p_term_deg, p.velocity_i_term_deg, p.pitch_error_deg, p.vel_error,
-            p.measured_vel_sps, applied_avg_sps, motor_dt_warning ? "  MOTOR_DT!" : "", p.turn_sps);
+            static_cast<double>(p.t_sec), static_cast<double>(p.pitch_deg),
+            static_cast<double>(p.pitch_rate_dps), static_cast<double>(p.u_sps),
+            (std::abs(p.u_sps) >= 0.99f * static_cast<float>(kMaxSps)) ? "*" : "",
+            static_cast<double>(p.pitch_sp_deg), static_cast<double>(p.velocity_p_term_deg),
+            static_cast<double>(p.velocity_i_term_deg), static_cast<double>(p.pitch_error_deg),
+            static_cast<double>(p.vel_error), static_cast<double>(p.measured_vel_sps),
+            static_cast<double>(applied_avg_sps), motor_dt_warning ? "  MOTOR_DT!" : "",
+            static_cast<double>(p.turn_sps));
       }
     }
   }
